api/requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
tensorflow==2.16.1